// angular-mat-table-crud-master/src/app/models/subscription.ts
export class Subscribers {
  subscription_id: string;
  tenant: string;
  client_id: string;
  secret: string;
  username: string;
  password: string;
  state:string;
  amount:number;
  environmentName:string;
  id:string;
  isDefault:string;
  name:string;
//  vm---------------
  // name: string;
  vm_name:string;
  type: string;
  status: string;
  resource_group: string;
  location: string;
  subscription: string;

  //vn
  resource_group_name:string;
}